// src/lib/pyno-datepicker.type.ts
export type PynoDatepickerDayOfWeek =
  | 'sunday'
  | 'monday'
  | 'tuesday'
  | 'wednesday'
  | 'thursday'
  | 'friday'
  | 'saturday';
export type PynoDatepickerDateObject = {
  gYear: string,
  gMonth: string,
  gDay: string,
  gDate: string,
  jYear: string,
  jMonth: string,
  jDay: string,
  jDate: string,
  timestamp: number,
  weekday: string
}